+++ lib/core/engine_templight.cpp
@@ -1,5 +1,5 @@
 // Metashell - Interactive C++ template metaprogramming shell
-// Copyright (C) 2015, Abel Sinkovics (abel@sinkovics.hu)
+// Copyright (C) 2018, Abel Sinkovics (abel@sinkovics.hu)
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -14,14 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/engine_clang.hpp>
+#include <metashell/engine_templight.hpp>
 
 #include <metashell/code_completer_clang.hpp>
 #include <metashell/cpp_validator_clang.hpp>
 #include <metashell/engine.hpp>
 #include <metashell/header_discoverer_clang.hpp>
 #include <metashell/macro_discovery_clang.hpp>
-#include <metashell/metaprogram_tracer_clang.hpp>
+#include <metashell/metaprogram_tracer_templight.hpp>
 #include <metashell/not_supported.hpp>
 #include <metashell/preprocessor_shell_clang.hpp>
 #include <metashell/type_shell_clang.hpp>
@@ -43,20 +43,22 @@ namespace
             data::feature::macro_discovery()};
   }
 
+  template <bool UseInternalTemplight>
   std::unique_ptr<iface::engine>
-  create_clang_engine(const data::config& config_,
-                      const boost::filesystem::path& internal_dir_,
-                      const boost::filesystem::path& temp_dir_,
-                      const boost::filesystem::path& env_filename_,
-                      iface::environment_detector& env_detector_,
-                      iface::displayer& displayer_,
-                      logger* logger_)
+  create_templight_engine(const data::config& config_,
+                          const boost::filesystem::path& internal_dir_,
+                          const boost::filesystem::path& temp_dir_,
+                          const boost::filesystem::path& env_filename_,
+                          iface::environment_detector& env_detector_,
+                          iface::displayer& displayer_,
+                          logger* logger_)
   {
     const clang_binary cbin(
-        false, find_clang(false, config_.active_shell_config().engine_args,
-                          config_.metashell_binary,
-                          config_.active_shell_config().engine, env_detector_,
-                          displayer_, logger_),
+        UseInternalTemplight,
+        find_clang(
+            UseInternalTemplight, config_.active_shell_config().engine_args,
+            config_.metashell_binary, config_.active_shell_config().engine,
+            env_detector_, displayer_, logger_),
         config_.active_shell_config().engine_args, internal_dir_, env_detector_,
         logger_);
 
@@ -66,23 +68,33 @@ namespace
         preprocessor_shell_clang(cbin),
         code_completer_clang(
             internal_dir_, temp_dir_, env_filename_, cbin, logger_),
-        header_discoverer_clang(cbin), metaprogram_tracer_clang(cbin),
+        header_discoverer_clang(cbin), metaprogram_tracer_templight(cbin),
         cpp_validator_clang(internal_dir_, env_filename_, cbin, logger_),
         macro_discovery_clang(cbin), not_supported(), supported_features());
   }
 } // anonymous namespace
 
-engine_entry metashell::get_engine_clang_entry()
+engine_entry metashell::get_engine_templight_entry()
 {
   return engine_entry(
-      &create_clang_engine,
-      "<Clang binary> -std=<standard to use> [<Clang args>]",
+      &create_templight_engine<false>,
+      "<Templight binary> -std=<standard to use> [<Clang args>]",
       data::markdown_string(
-          "Uses the [Clang compiler](http://clang.llvm.org). `<Clang args>` "
-          "are passed to the compiler as command line-arguments. Note that "
-          "Metashell requires C++11 or above. If your Clang uses such a "
-          "standard by default, you can omit the `-std` argument. Metaprogram "
-          "debugging (MDB) is supported only when Clang has been patched with "
-          "[templight](https://github.com/mikael-s-persson/templight)"),
+          "Uses [Templight](https://github.com/mikael-s-persson/templight). "
+          "`<Clang args>` are passed to the compiler as command "
+          "line-arguments. Note that Metashell requires C++11 or above. If "
+          "your Templight uses such a standard by default, you can omit the "
+          "`-std` argument."),
+      supported_features());
+}
+
+engine_entry metashell::get_internal_templight_entry()
+{
+  return engine_entry(
+      &create_templight_engine<true>, "[<Clang args>]",
+      data::markdown_string(
+          "Uses the [Templight](https://github.com/mikael-s-persson/templight) "
+          "shipped with Metashell. `<Clang args>` are passed to"
+          " the compiler as command line-arguments."),
       supported_features());
 }